# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "课程智能体系统 API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/graph": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Graph",
        "description": "三张图谱共用入口。\n\ntype=knowledge 知识图谱（节点值=掌握度）\ntype=problem   问题图谱（节点值=错误率，来自真实作答统计）\ntype=goal      目标图谱（节点值=达成度，按关联知识点加权）\n\nuserId 存在时按该学生视角计算，否则按当前用户所属班级计算。",
        "operationId": "graph_api_v1_graph_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "knowledge",
              "title": "Type"
            }
          },
          {
            "name": "courseId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "C2026DS001",
              "title": "Courseid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Userid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/kp/{kp_id}": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Kp Detail",
        "description": "知识点详情：前测/后测掌握度、关联资源、题量、错题量。",
        "operationId": "kp_detail_api_v1_graph_kp__kp_id__get",
        "parameters": [
          {
            "name": "kp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kp Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/path": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Learning Path",
        "description": "学习路径：节点状态（done/todo/warn/doing）由真实掌握度推导。",
        "operationId": "learning_path_api_v1_graph_path_get",
        "parameters": [
          {
            "name": "courseId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "C2026DS001",
              "title": "Courseid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/intervention/list": {
      "get": {
        "tags": [
          "intervention"
        ],
        "summary": "List Interventions",
        "operationId": "list_interventions_api_v1_intervention_list_get",
        "parameters": [
          {
            "name": "classId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Classid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Scope"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intervention/{iv_id}/confirm": {
      "post": {
        "tags": [
          "intervention"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_v1_intervention__iv_id__confirm_post",
        "parameters": [
          {
            "name": "iv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iv Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intervention/{iv_id}/reject": {
      "post": {
        "tags": [
          "intervention"
        ],
        "summary": "Reject",
        "operationId": "reject_api_v1_intervention__iv_id__reject_post",
        "parameters": [
          {
            "name": "iv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iv Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intervention/{iv_id}/effect": {
      "get": {
        "tags": [
          "intervention"
        ],
        "summary": "Effect",
        "description": "干预效果：干预组 / 对照组正确率曲线 + 净提升（真实作答统计）。",
        "operationId": "effect_api_v1_intervention__iv_id__effect_get",
        "parameters": [
          {
            "name": "iv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iv Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intervention/templates": {
      "get": {
        "tags": [
          "intervention"
        ],
        "summary": "Templates",
        "operationId": "templates_api_v1_intervention_templates_get",
        "parameters": [
          {
            "name": "scene",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Scene"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Keyword"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "intervention"
        ],
        "summary": "Save Template",
        "description": "策略沉淀：真实落 strategy_templates 表。",
        "operationId": "save_template_api_v1_intervention_templates_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "title": "Data"
          },
          "traceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traceid"
          }
        },
        "type": "object",
        "title": "Envelope"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}